Attempted to implement scaling with pygame, but... well, I failed miserably.
As of right now, the current folder (with version 2.2) does work, but scaling is 50/50. Some works, some is broken. However, the very fundamentals of the program are still intact and work. Just... try not to click the magnifying glass.

Next version, I will try using tkinter instead of pygame for handling to graphics. Hopefully, things will work out better. Hopefully... ;)